:recycle: Move const variable to const file
