+++ src/renderer/const.ts
@@ -21,3 +21,20 @@ export const BUILD_FOR_TRAVIS_COMMIT = __BUILD_FOR_TRAVIS_COMMIT__;
 
 // @ts-ignore
 export const APP_VERSION = __PACKAGE_INFO_VERSION__;
+
+export const WHITELIST_DOMAINS = [
+    'alilang-desktop-client.cn-hangzhou.log.aliyuncs.com',
+    's-api.alibaba-inc.com',
+    'alilang.alibaba-inc.com',
+    'auth-alilang.alibaba-inc.com',
+    'mdm-alilang.alibaba-inc.com',
+    '***.apple.com',
+    '*.mzstatic.com',
+    '*.cdn-apple.com', 
+    '***.apple-cloudkit.com',
+    '***.icloud.com',
+    '***.icloud-content.com',
+    '***.icloud.com.cn',
+    'txy.live-play.acgvideo.com',
+    'ocs-oneagent-server.alibaba.com',
+];
